requirements.txt
aiohttp==3.13.3
colorlog==6.10.1
pywinrm==0.5.0
